Extending The User Model
When you have to deal with user accounts, you will find that the user
model of the Django authentication framework is suitable for common
cases. However, the user model comes with very basic fields. You may
wish to extend it to include additional data. The best way to do this
is by creating a profile model that contains all additional fields
and a one-to-one relationship with the Django User model. A one-to-one
relationship is similar to a ForeignKey field with the parameter
unique=True. The reverse side of the relationship is an implicit
one-to-one relationship with the related model instead of a manager for
multiple elements. From each side of the relationship, you retrieve a
single related object.